+++ ta4j-core/src/main/java/org/ta4j/core/backtest/BacktestExecutor.java
@@ -21,11 +21,15 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core;
+package org.ta4j.core.backtest;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
+import org.ta4j.core.BarSeries;
+import org.ta4j.core.Strategy;
+import org.ta4j.core.Trade;
+import org.ta4j.core.TradingRecord;
 import org.ta4j.core.analysis.cost.CostModel;
 import org.ta4j.core.analysis.cost.ZeroCostModel;
 import org.ta4j.core.num.Num;
@@ -45,17 +49,18 @@ public class BacktestExecutor {
         this(series, new TradingStatementGenerator());
     }
 
-    public BacktestExecutor(BarSeries series, CostModel transactionCostModel, CostModel holdingCostModel) {
-        this(series, new TradingStatementGenerator(), transactionCostModel, holdingCostModel);
+    public BacktestExecutor(BarSeries series, CostModel transactionCostModel, CostModel holdingCostModel,
+            TradeExecutionModel tradeExecutionModel) {
+        this(series, new TradingStatementGenerator(), transactionCostModel, holdingCostModel, tradeExecutionModel);
     }
 
     public BacktestExecutor(BarSeries series, TradingStatementGenerator tradingStatementGenerator) {
-        this(series, tradingStatementGenerator, new ZeroCostModel(), new ZeroCostModel());
+        this(series, tradingStatementGenerator, new ZeroCostModel(), new ZeroCostModel(), new TradeOnNextOpenModel());
     }
 
     public BacktestExecutor(BarSeries series, TradingStatementGenerator tradingStatementGenerator,
-            CostModel transactionCostModel, CostModel holdingCostModel) {
-        this.seriesManager = new BarSeriesManager(series, transactionCostModel, holdingCostModel);
+            CostModel transactionCostModel, CostModel holdingCostModel, TradeExecutionModel tradeExecutionModel) {
+        this.seriesManager = new BarSeriesManager(series, transactionCostModel, holdingCostModel, tradeExecutionModel);
         this.tradingStatementGenerator = tradingStatementGenerator;
     }