+++ ta4j-core/src/test/java/org/ta4j/core/backtest/BarSeriesManagerTest.java
@@ -21,11 +21,12 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core;
+package org.ta4j.core.backtest;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
+import java.time.Duration;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
@@ -34,9 +35,19 @@ import java.util.function.Function;
 
 import org.junit.Before;
 import org.junit.Test;
+import org.ta4j.core.BaseBar;
+import org.ta4j.core.BaseBarBuilder;
+import org.ta4j.core.BarSeries;
+import org.ta4j.core.BaseBarSeries;
+import org.ta4j.core.BaseStrategy;
+import org.ta4j.core.Position;
+import org.ta4j.core.Strategy;
+import org.ta4j.core.Trade;
 import org.ta4j.core.Trade.TradeType;
+import org.ta4j.core.TradingRecord;
 import org.ta4j.core.indicators.AbstractIndicatorTest;
 import org.ta4j.core.mocks.MockBarSeries;
+import org.ta4j.core.num.DecimalNum;
 import org.ta4j.core.num.Num;
 import org.ta4j.core.rules.FixedRule;
 
@@ -50,6 +61,8 @@ public class BarSeriesManagerTest extends AbstractIndicatorTest<BarSeries, Num> 
 
     private final Num HUNDRED = numOf(100);
 
+    private final TradeExecutionModel tradeExecutionModel = new TradeOnCurrentCloseModel();
+
     public BarSeriesManagerTest(Function<Number, Num> numFunction) {
         super(numFunction);
     }
@@ -58,6 +71,7 @@ public class BarSeriesManagerTest extends AbstractIndicatorTest<BarSeries, Num> 
     public void setUp() {
 
         final DateTimeFormatter dtf = DateTimeFormatter.ISO_ZONED_DATE_TIME;
+
         seriesForRun = new MockBarSeries(numFunction, new double[] { 1d, 2d, 3d, 4d, 5d, 6d, 7d, 8d, 9d },
                 new ZonedDateTime[] { ZonedDateTime.parse("2013-01-01T00:00:00-05:00", dtf),
                         ZonedDateTime.parse("2013-08-01T00:00:00-05:00", dtf),
@@ -68,7 +82,7 @@ public class BarSeriesManagerTest extends AbstractIndicatorTest<BarSeries, Num> 
                         ZonedDateTime.parse("2015-08-01T00:00:00-05:00", dtf),
                         ZonedDateTime.parse("2015-10-01T00:00:00-05:00", dtf),
                         ZonedDateTime.parse("2015-12-01T00:00:00-05:00", dtf) });
-        manager = new BarSeriesManager(seriesForRun);
+        manager = new BarSeriesManager(seriesForRun, tradeExecutionModel);
 
         strategy = new BaseStrategy(new FixedRule(0, 2, 3, 6), new FixedRule(1, 4, 7, 8));
         strategy.setUnstableBars(2); // Strategy would need a real test class
@@ -163,7 +177,7 @@ public class BarSeriesManagerTest extends AbstractIndicatorTest<BarSeries, Num> 
                         dateTime.withYear(2001), dateTime.withYear(2002), dateTime.withYear(2002),
                         dateTime.withYear(2002), dateTime.withYear(2003), dateTime.withYear(2004),
                         dateTime.withYear(2005) });
-        manager = new BarSeriesManager(series);
+        manager = new BarSeriesManager(series, tradeExecutionModel);
 
         Strategy aStrategy = new BaseStrategy(new FixedRule(0, 3, 5, 7), new FixedRule(2, 4, 6, 9));
 
@@ -198,5 +212,63 @@ public class BarSeriesManagerTest extends AbstractIndicatorTest<BarSeries, Num> 
         // no trade happened within [9-9]
         positions = manager.run(aStrategy, 9, 9).getPositions();
         assertTrue(positions.isEmpty());
+    }
+
+    @Test
+    public void runOnSeriesSlicesUsingExecuteOnNextBarOpenPriceModel() {
+        ZonedDateTime dateTime = ZonedDateTime.of(2000, 1, 1, 0, 0, 0, 0, ZoneId.systemDefault());
+        BarSeries series = new BaseBarSeries();
+        BaseBarBuilder builder = BaseBar.builder();
+        Duration duration = Duration.ofDays(1);
+        Num one = DecimalNum.valueOf(1);
+        Num two = DecimalNum.valueOf(2);
+        Num three = DecimalNum.valueOf(3);
+        Num four = DecimalNum.valueOf(4);
+        series.addBar(builder.timePeriod(duration)
+                .endTime(dateTime.withYear(2000))
+                .openPrice(one)
+                .highPrice(one)
+                .lowPrice(one)
+                .closePrice(one)
+                .build());
+        series.addBar(builder.timePeriod(duration)
+                .endTime(dateTime.withYear(2001))
+                .openPrice(two)
+                .highPrice(two)
+                .lowPrice(two)
+                .closePrice(two)
+                .build());
+        series.addBar(builder.timePeriod(duration)
+                .endTime(dateTime.withYear(2002))
+                .openPrice(three)
+                .highPrice(three)
+                .lowPrice(three)
+                .closePrice(three)
+                .build());
+        series.addBar(builder.timePeriod(duration)
+                .endTime(dateTime.withYear(2003))
+                .openPrice(four)
+                .highPrice(four)
+                .lowPrice(four)
+                .closePrice(four)
+                .build());
+        manager = new BarSeriesManager(series, new TradeOnNextOpenModel());
+
+        Strategy aStrategy = new BaseStrategy(new FixedRule(0, 3, 5, 7), new FixedRule(2, 4, 6, 9));
+
+        // there are 0 entries within [0-1], the entry was analysed during bar 0, but
+        // executed at the open of bar 1
+        TradingRecord tradingRecord = manager.run(aStrategy, 0, 1);
+        List<Position> positions = tradingRecord.getPositions();
+        assertEquals(0, positions.size());
+
+        // only 1 entry happened within [0-2]
+        tradingRecord = manager.run(aStrategy, 0, 3);
+        positions = tradingRecord.getPositions();
+        assertEquals(1, positions.size());
+        assertEquals(1, positions.get(0).getEntry().getIndex());
+        // Entry price should be 2 as we open on the open of the next bar that produced
+        // the entry signal
+        assertEquals(two, positions.get(0).getEntry().getNetPrice());
     }
 }